Revert "implementingMaster"
