+++ 03-D3D/D3DDemo.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2013
+VisualStudioVersion = 12.0.30110.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "BasicWindow", "Demo.vcxproj", "{9AFF6E4E-5D21-4DDE-948B-0E6BE0365098}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Win32 = Debug|Win32
+		Release|Win32 = Release|Win32
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9AFF6E4E-5D21-4DDE-948B-0E6BE0365098}.Debug|Win32.ActiveCfg = Debug|Win32
+		{9AFF6E4E-5D21-4DDE-948B-0E6BE0365098}.Debug|Win32.Build.0 = Debug|Win32
+		{9AFF6E4E-5D21-4DDE-948B-0E6BE0365098}.Release|Win32.ActiveCfg = Release|Win32
+		{9AFF6E4E-5D21-4DDE-948B-0E6BE0365098}.Release|Win32.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
