--- 03-D3D/Projekt med klasser/Engine.cpp
@@ -1,239 +1,0 @@
-#include "Linker.h"
-#include "Engine.h"
-
-Engine::Engine()
-{
-}
-
-Engine::~Engine()
-{
-	gSwapChain->Release();
-	gDevice->Release();
-	gDeviceContext->Release();
-	gBackbufferRTV->Release();
-
-	gDepthview->Release();
-	gDepthStencilBuffer->Release();
-
-	gVertexLayout->Release();
-	gVertexShader->Release();
-	gConstantBuffer->Release();
-	gPixelShader->Release();
-}
-
-ID3D11Device* Engine::GetDevice()
-{
-	return gDevice;
-}
-
-ID3D11DeviceContext* Engine::GetDeviceContext()
-{
-	return gDeviceContext;
-}
-
-void Engine::CreateShaders()
-{
-	//create vertex shader
-	ID3DBlob* pVS = nullptr;
-	D3DCompileFromFile(
-		L"VertexShader.hlsl", // filename
-		nullptr,		// optional macros
-		nullptr,		// optional include files
-		"VS_main",		// entry point
-		"vs_4_0",		// shader model (target)
-		0,				// shader compile options
-		0,				// effect compile options
-		&pVS,			// double pointer to ID3DBlob		
-		nullptr			// pointer for Error Blob messages.			
-		);
-
-	D3D11_INPUT_ELEMENT_DESC inputDesc[] = {
-		{ "POSITION", 0, DXGI_FORMAT_R32G32B32_FLOAT, 0, 0, D3D11_INPUT_PER_VERTEX_DATA, 0 },
-		{ "NORMAL", 0, DXGI_FORMAT_R32G32B32_FLOAT, 0, 12, D3D11_INPUT_PER_VERTEX_DATA, 0 },
-		{ "TEXCOORD", 0, DXGI_FORMAT_R32G32_FLOAT, 0, 24, D3D11_INPUT_PER_VERTEX_DATA, 0 },
-	};
-	gDevice->CreateInputLayout(inputDesc, ARRAYSIZE(inputDesc), pVS->GetBufferPointer(), pVS->GetBufferSize(), &gVertexLayout);
-	// we do not need anymore this COM object, so we release it.
-	pVS->Release();
-
-	gDevice->CreateVertexShader(pVS->GetBufferPointer(), pVS->GetBufferSize(), nullptr, &gVertexShader);
-
-	//create pixel shader
-	ID3DBlob* pPS = nullptr;
-	D3DCompileFromFile(
-		L"PixelShader.hlsl", // filename
-		nullptr,		// optional macros
-		nullptr,		// optional include files
-		"PS_main",		// entry point
-		"ps_4_0",		// shader model (target)
-		0,				// shader compile options
-		0,				// effect compile options
-		&pPS,			// double pointer to ID3DBlob		
-		nullptr			// pointer for Error Blob messages.
-		);
-
-	gDevice->CreatePixelShader(pPS->GetBufferPointer(), pPS->GetBufferSize(), nullptr, &gPixelShader);
-	// we do not need anymore this COM object, so we release it.
-	pPS->Release();
-
-	/*//create geometry shader
-	ID3DBlob* pGS = nullptr;
-	D3DCompileFromFile(
-		L"GeometryShader.hlsl", // filename
-		nullptr,		// optional macros
-		nullptr,		// optional include files
-		"GS_main",		// entry point
-		"gs_4_0",		// shader model (target)
-		0,				// shader compile options
-		0,				// effect compile options
-		&pGS,			// double pointer to ID3DBlob		
-		nullptr			// pointer for Error Blob messages.
-		);
-
-	gDevice->CreateGeometryShader(pGS->GetBufferPointer(), pGS->GetBufferSize(), nullptr, &gGeometryShader);
-	pGS->Release();
-	*/
-}
-
-void Engine::SetViewPort()
-{
-	D3D11_VIEWPORT vp;
-	vp.Width = (float)1280;
-	vp.Height = (float)960;
-	vp.MinDepth = 0.0f;
-	vp.MaxDepth = 1.0f;
-	vp.TopLeftX = 0;
-	vp.TopLeftY = 0;
-	gDeviceContext->RSSetViewports(1, &vp);
-}
-
-void Engine::Render()
-{
-	float clearColor[] = { 0, 0, 0, 1 };
-	
-	gDeviceContext->ClearRenderTargetView(gBackbufferRTV, clearColor);
-	gDeviceContext->ClearDepthStencilView(gDepthview, D3D11_CLEAR_DEPTH, 1.0f, 0);
-
-	gDeviceContext->VSSetShader(gVertexShader, nullptr, 0);
-	gDeviceContext->HSSetShader(nullptr, nullptr, 0);
-	gDeviceContext->DSSetShader(nullptr, nullptr, 0);
-	gDeviceContext->GSSetShader(nullptr, nullptr, 0);
-	gDeviceContext->PSSetShader(gPixelShader, nullptr, 0);
-}
-
-HRESULT Engine::CreateDirect3DContext(HWND wndHandle)
-{
-	// create a struct to hold information about the swap chain
-	DXGI_SWAP_CHAIN_DESC scd;
-
-	// clear out the struct for use
-	ZeroMemory(&scd, sizeof(DXGI_SWAP_CHAIN_DESC));
-
-	// fill the swap chain description struct
-	scd.BufferCount = 1;                                    // one back buffer
-	scd.BufferDesc.Format = DXGI_FORMAT_R8G8B8A8_UNORM;     // use 32-bit color
-	scd.BufferUsage = DXGI_USAGE_RENDER_TARGET_OUTPUT;      // how swap chain is to be used
-	scd.OutputWindow = wndHandle;                           // the window to be used
-	scd.SampleDesc.Count = 4;                               // how many multisamples
-	scd.Windowed = TRUE;                                    // windowed/full-screen mode
-	scd.BufferDesc.RefreshRate.Numerator = 60;				//Set the refresh rate to 60. 
-	// create a device, device context and swap chain using the information in the scd struct
-	
-	HRESULT hr = D3D11CreateDeviceAndSwapChain(NULL,
-		D3D_DRIVER_TYPE_HARDWARE,
-		NULL,
-		NULL,
-		NULL,
-		NULL,
-		D3D11_SDK_VERSION,
-		&scd,
-		&gSwapChain,
-		&gDevice,
-		NULL,
-		&gDeviceContext);
-
-	if (SUCCEEDED(hr))
-	{
-		// get the address of the back buffer
-		ID3D11Texture2D* pBackBuffer = nullptr;
-		gSwapChain->GetBuffer(0, __uuidof(ID3D11Texture2D), (LPVOID*)&pBackBuffer);
-
-		CreateDepthBuffer(); //It's very important to call the creation of the Depth Buffer here!
-
-		// use the back buffer address to create the render target
-		gDevice->CreateRenderTargetView(pBackBuffer, NULL, &gBackbufferRTV);
-		pBackBuffer->Release();
-		
-		// set the render target as the back buffer
-		gDeviceContext->OMSetRenderTargets(1, &gBackbufferRTV, gDepthview);
-	}
-	return hr;
-}
-
-void Engine::CreateDepthBuffer()
-{
-	D3D11_TEXTURE2D_DESC desc;
-	desc.Width = 1280;
-	desc.Height = 960;
-	desc.MipLevels = 1;
-	desc.ArraySize = 1;
-	desc.Format = DXGI_FORMAT_D32_FLOAT;
-	desc.SampleDesc.Count = 4;
-	desc.SampleDesc.Quality = 0;
-	desc.Usage = D3D11_USAGE_DEFAULT;
-	desc.BindFlags = D3D11_BIND_DEPTH_STENCIL;
-	desc.CPUAccessFlags = 0;
-	desc.MiscFlags = 0;
-
-	gDevice->CreateTexture2D(&desc, 0, &gDepthStencilBuffer);
-
-	gDevice->CreateDepthStencilView(gDepthStencilBuffer, 0, &gDepthview);
-}
-
-void Engine::CreateConstantBuffer()
-{
-	D3D11_BUFFER_DESC desc;
-	desc.Usage = D3D11_USAGE_DYNAMIC;
-	desc.ByteWidth = sizeof(VS_CONSTANT_BUFFER);
-	desc.BindFlags = D3D11_BIND_CONSTANT_BUFFER;
-	desc.CPUAccessFlags = D3D11_CPU_ACCESS_WRITE;
-	desc.MiscFlags = 0;
-	desc.StructureByteStride = 0;
-
-	//Creating the Constant Buffer
-	gDevice->CreateBuffer(&desc, NULL, &gConstantBuffer);
-}
-
-void Engine::UpdateConstantBuffer()
-{
-	HRESULT result;
-	D3D11_MAPPED_SUBRESOURCE mappedCB;
-	VS_CONSTANT_BUFFER* dataPtr;
-
-	Matrix world;
-	Matrix projection;
-	Matrix camView;
-	Matrix worldViewProjection;
-
-	world = XMMatrixTranslation(0, 0, 0);
-	projection = XMMatrixPerspectiveFovLH(float(3.1415*0.45), float(1280.0 / 960.0), float(0.5), float(20));
-
-	worldViewProjection = world * cameraPtr->camView * projection;
-	worldViewProjection = worldViewProjection.Transpose();
-
-	world = world.Transpose();
-
-	result = gDeviceContext->Map(gConstantBuffer, 0, D3D11_MAP_WRITE_DISCARD, 0, &mappedCB);
-	if (FAILED(result))
-	{
-		return;
-	}
-
-	dataPtr = (VS_CONSTANT_BUFFER*)mappedCB.pData;
-	dataPtr->worldViewProj = worldViewProjection;
-	dataPtr->world = world;
-
-	gDeviceContext->Unmap(gConstantBuffer, 0);
-
-	gDeviceContext->VSSetConstantBuffers(0, 1, &gConstantBuffer);
-}
