--- 03-D3D/Projekt med klasser/bth_image.h
@@ -1,118 +1,0 @@
-/*
-AN EARLY CHRISTMAS GIFT FROM THE TEACHERS
-Ha kul oenskar Stefan, Francisco, Tim, Kim och Alexander :-)
-
-
-                           .,;'.:0d'xN0c,oNNOk;.'lc.                            
-                       .:,.;0WXkkKOckN0l:xNNko'.kNNkod:.                        
-                   .lkoOO, .xWWNWWXkONKx::d0WNolNMWXo'.   .;,.                  
-               .,c''OWWWXo'.lX0:o0KkoxxllodxOdlkXNWNo  .;xKWXc                  
-             .cOXd:;:OWMWNKdc:;                .,,lkd:lOKXNWX:                  
-           ,;':0WN0l,lO0o'..                        ..;;.:KWk'                  
-         'dXk' 'xNN0dc'                                  .lkc.    .lko.         
-        :kONN0c..:l;.                                           ,dKWXd,..       
-        ...,dXNk'                  ....   ....                 ,0X0O0xkKx'      
-             .:'                  cXWNKxcdNWNx.                 '';dXWNOc.      
-   .,'                            ;0NNNXxkNNXo.                   ;OKkddxxo,''  
-  .oOocc...            ,cl:'':cc:',x00kl;cO0Oc',clc,.;cl:.         .'oKNKkdkkd: 
-  ;ONNNOddo.          .dWWW0kKWWW0ONWWNKdkWWNxkNWWW0ONWWXl          .oOc..'xOo' 
- .'.,dOXXd.      .'....lKNXxlxXNKxoOXNKkldXNKll0NN0dd0NNk, ..'.      'dkk0XXx'  
-'xl.'OKXl.       xNNN0lxNNXxd0NNKxxXNNKkcdXNXxd0NNNkdKNNKxdKNNk'       ':odo:'. 
-lOl..ck0:        lXWWNdxNWW0xkNWW0kXWWWXxOWWNdxNWWNOONWW0xONWWk.        ,ONNNXx,
-,0NXKK0l.        cOOOo;cOOkc:oOOxcckOOd:':kOkc;oOOOl:dOOd::xOOo.       .o0d:';dx
-.,lxkd:.         xWMMNxOWMW0kKWMWKOWMMWKdkWMWxkNMMW0OWMWXOONMMO'       .c0Ox,.cx
-O00KK0o.         :kKKkclOK0xco0KOdlk0K0xlo0KOclOKKOooOK0xldOK0o.        .x0x, . 
-cxXWW0c.   ;x00x:dKXKkcoKXKdlkKX0do0KKOo;oKX0olkKXKdoOKKOoo0XKx:dO00x.  .xd;ckko
-0NMW0o,    ;0WWXdONWWNxkWWWKk0WWW0ONWWWXxOWWNxkWWWN0ONWWKk0NWW0dXWWNk.  ,0xOWXOX
-kkOO00x.   .oOOx;:xOOx;:kOOo:lkOkl:dOOko:lOOk::xOOxcckOOo:lkOkl;dOOx:   .d0KOl:x
-:dxkOK0;   cKMMXo0WMWXdkWMWOkKWMWOkWMMNOoxWMWkxXWMW0kNWWXkkNMW0oKWMM0.  ;k0ko;'.
-oOOkdddc.  'kXNKoo0NNXodKNXOdxKNNOdOXNNKxkNNKldXNX0xxXNXkokXNXxl0NX0l. .dNMMWNKd
-.'.  ,k0:   .....cOK0x;cO0Oocd00Odx0NNNNXXNNXkox00klck00xclOK0o'....   :KXKKXKl.
-.dkxxXWNd.       xWMMNxOWMWKOKWMWWWMMMMMMMMMWXKWMMW00WMWXO0NMWO'      .:dc..:kl.
- .oXWWWNKo.      .cddo;;lddl,;oddkXMMMMMMMMMWKkddoc;:odo:,codo,      ,dO0KKx;.. 
-  ;docc:,lx,         ............;OWMMMMMMMMMNl.............        '0d'.;dXd.  
-   .;;:kOON0;       'k00000000000XWWWWWWWWWWWWNK00000000000O'      .;KNkl;cx:   
-    :OKNWKxodl.     .:lllllllllllllllllllllllllllllllllllllc.    .dkoxOKK0d,    
-     'do;. .c00:                                               .:x0XNKkl,.      
-         .;xXWXx:'.                                           ,do'..l0N0:       
-        .lXNOclKXKkl'                                      ...lXk.   .,'        
-         .;,'l0NNKXWO,                                   .lKX0KWNc              
-            .dXKNWNk,                                     .;o0NWWO'             
-             .;dK0c.  ,do.   ,lc;,. .........'..;oo'   cxc   .'cl,              
-                ..    ,do.  .xNN0KKo;ckKKx:c0Nx;dNNl  .ckc.                     
-                            ;KWXXW0; .xWWo .kWNOkNWk.                           
-                            cXXkKNO, .xWWl  cKWd.dKO;                           
-                            .;codo;  .xWWl  'ONo..'.    
-
-*/
-
-const unsigned int BTH_IMAGE_WIDTH = 64;
-const unsigned int BTH_IMAGE_HEIGHT = 64;
-
-//Image data stored in 8-bit RGBA format
-unsigned char BTH_IMAGE_DATA[] = {
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,16,16,16,16,0,0,0,0,143,143,143,143,223,223,223,223,0,0,0,0,191,191,191,191,255,255,255,255,128,128,128,128,16,16,16,16,207,207,207,207,255,255,255,255,255,255,255,255,191,191,191,191,16,16,16,16,32,32,32,32,80,80,80,80,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,0,0,0,0,159,159,159,159,255,255,255,255,207,207,207,207,48,48,48,48,159,159,159,159,255,255,255,255,32,32,32,32,191,191,191,191,255,255,255,255,128,128,128,128,128,128,128,128,255,255,255,255,223,223,223,223,32,32,32,32,175,175,175,175,0,0,0,0,112,112,112,112,255,255,255,255,255,255,255,255,0,0,0,0,80,80,80,80,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,159,159,159,159,175,175,175,175,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,239,239,239,239,159,159,159,159,255,255,255,255,96,96,96,96,191,191,191,191,255,255,255,255,128,128,128,128,96,96,96,96,255,255,255,255,255,255,255,255,239,239,239,239,80,80,80,80,0,0,0,0,207,207,207,207,255,255,255,255,207,207,207,207,207,207,207,207,255,255,255,255,191,191,191,191,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,175,175,175,175,64,64,64,64,255,255,255,255,96,96,96,96,0,0,0,0,80,80,80,80,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,143,143,143,143,191,191,191,191,255,255,255,255,159,159,159,159,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,239,239,239,239,32,32,32,32,255,255,255,255,255,255,255,255,255,255,255,255,223,223,223,223,80,80,80,80,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,255,255,255,255,255,255,255,255,207,207,207,207,255,255,255,255,16,16,16,16,0,0,0,0,16,16,16,16,255,255,255,255,223,223,223,223,207,207,207,207,255,255,255,255,255,255,255,255,207,207,207,207,175,175,175,175,255,255,255,255,191,191,191,191,175,175,175,175,112,112,112,112,80,80,80,80,255,255,255,255,255,255,255,255,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,143,143,143,143,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,80,80,80,80,223,223,223,223,0,0,0,0,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,32,32,32,32,0,0,0,0,223,223,223,223,223,223,223,223,16,16,16,16,191,191,191,191,255,255,255,255,255,255,255,255,128,128,128,128,223,223,223,223,143,143,143,143,143,143,143,143,239,239,239,239,255,255,255,255,239,239,239,239,96,96,96,96,175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,223,223,223,223,0,0,0,0,0,0,0,0,16,16,16,16,175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,143,143,143,143,239,239,239,239,96,96,96,96,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,191,191,191,191,223,223,223,223,0,0,0,0,0,0,0,0,48,48,48,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,96,96,96,96,159,159,159,159,128,128,128,128,255,255,255,255,255,255,255,255,48,48,48,48,48,48,48,48,207,207,207,207,255,255,255,255,207,207,207,207,255,255,255,255,223,223,223,223,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,159,159,159,159,255,255,255,255,239,239,239,239,96,96,96,96,207,207,207,207,80,80,80,80,191,191,191,191,255,255,255,255,255,255,255,255,223,223,223,223,239,239,239,239,143,143,143,143,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,80,80,80,80,143,143,143,143,159,159,159,159,239,239,239,239,255,255,255,255,207,207,207,207,255,255,255,255,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,80,80,80,80,48,48,48,48,239,239,239,239,255,255,255,255,255,255,255,255,159,159,159,159,0,0,0,0,64,64,64,64,255,255,255,255,239,239,239,239,48,48,48,48,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,48,48,48,48,0,0,0,0,191,191,191,191,255,255,255,255,128,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,239,239,239,239,143,143,143,143,0,0,0,0,96,96,96,96,255,255,255,255,255,255,255,255,159,159,159,159,128,128,128,128,239,239,239,239,96,96,96,96,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,239,239,239,239,255,255,255,255,96,96,96,96,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,80,80,80,80,48,48,48,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,207,207,207,207,255,255,255,255,96,96,96,96,0,0,0,0,0,0,0,0,159,159,159,159,255,255,255,255,255,255,255,255,223,223,223,223,48,48,48,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,207,207,207,207,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,207,207,207,207,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,207,207,207,207,255,255,255,255,255,255,255,255,255,255,255,255,96,96,96,96,0,0,0,0,16,16,16,16,207,207,207,207,175,175,175,175,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,175,175,175,175,255,255,255,255,255,255,255,255,223,223,223,223,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,80,80,80,80,191,191,191,191,64,64,64,64,239,239,239,239,255,255,255,255,255,255,255,255,128,128,128,128,0,0,0,0,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,207,207,207,207,255,255,255,255,255,255,255,255,223,223,223,223,16,16,16,16,64,64,64,64,128,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,207,207,207,207,255,255,255,255,255,255,255,255,96,96,96,96,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,96,96,96,96,255,255,255,255,255,255,255,255,175,175,175,175,239,239,239,239,159,159,159,159,255,255,255,255,255,255,255,255,96,96,96,96,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,207,207,207,207,159,159,159,159,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,255,255,255,255,96,96,96,96,32,32,32,32,223,223,223,223,255,255,255,255,255,255,255,255,128,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,159,159,159,159,96,96,96,96,0,0,0,0,191,191,191,191,255,255,255,255,255,255,255,255,239,239,239,239,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,207,207,207,207,255,255,255,255,255,255,255,255,207,207,207,207,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,80,80,80,80,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,239,239,239,239,255,255,255,255,175,175,175,175,64,64,64,64,159,159,159,159,191,191,191,191,128,128,128,128,0,0,0,0,64,64,64,64,16,16,16,16,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,191,191,191,191,128,128,128,128,32,32,32,32,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,112,112,112,112,191,191,191,191,191,191,191,191,191,191,191,191,96,96,96,96,191,191,191,191,191,191,191,191,175,175,175,175,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,128,128,128,128,64,64,64,64,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,239,239,239,239,48,48,48,48,0,0,0,0,
-	0,0,0,0,0,0,0,0,48,48,48,48,255,255,255,255,96,96,96,96,159,159,159,159,159,159,159,159,0,0,0,0,112,112,112,112,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,191,191,191,191,191,191,191,191,175,175,175,175,48,48,48,48,143,143,143,143,191,191,191,191,191,191,191,191,143,143,143,143,48,48,48,48,191,191,191,191,191,191,191,191,191,191,191,191,80,80,80,80,16,16,16,16,159,159,159,159,191,191,191,191,191,191,191,191,96,96,96,96,80,80,80,80,191,191,191,191,191,191,191,191,191,191,191,191,48,48,48,48,143,143,143,143,191,191,191,191,191,191,191,191,143,143,143,143,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,191,191,191,191,255,255,255,255,239,239,239,239,159,159,159,159,96,96,96,96,223,223,223,223,112,112,112,112,175,175,175,175,0,0,0,0,
-	0,0,0,0,0,0,0,0,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,96,96,96,96,64,64,64,64,239,239,239,239,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,239,239,239,239,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,239,239,239,239,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,191,191,191,191,207,207,207,207,32,32,32,32,0,0,0,0,0,0,0,0,159,159,159,159,175,175,175,175,112,112,112,112,0,0,0,0,
-	0,0,0,0,0,0,0,0,128,128,128,128,223,223,223,223,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,159,159,159,159,143,143,143,143,0,0,0,0,16,16,16,16,143,143,143,143,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,96,96,96,96,223,223,223,223,255,255,255,255,239,239,239,239,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,191,191,191,191,255,255,255,255,255,255,255,255,191,191,191,191,64,64,64,64,239,239,239,239,255,255,255,255,255,255,255,255,128,128,128,128,143,143,143,143,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,223,223,223,223,32,32,32,32,255,255,255,255,255,255,255,255,255,255,255,255,143,143,143,143,128,128,128,128,255,255,255,255,255,255,255,255,239,239,239,239,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,0,0,0,0,0,0,0,0,
-	0,0,0,0,159,159,159,159,112,112,112,112,0,0,0,0,128,128,128,128,112,112,112,112,159,159,159,159,128,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,191,191,191,191,191,191,191,191,175,175,175,175,64,64,64,64,48,48,48,48,191,191,191,191,191,191,191,191,143,143,143,143,16,16,16,16,143,143,143,143,191,191,191,191,191,191,191,191,96,96,96,96,48,48,48,48,191,191,191,191,191,191,191,191,159,159,159,159,48,48,48,48,0,0,0,0,128,128,128,128,191,191,191,191,191,191,191,191,96,96,96,96,48,48,48,48,175,175,175,175,191,191,191,191,191,191,191,191,48,48,48,48,96,96,96,96,191,191,191,191,191,191,191,191,143,143,143,143,16,16,16,16,159,159,159,159,191,191,191,191,191,191,191,191,48,48,48,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,16,16,16,16,0,0,0,0,0,0,0,0,
-	32,32,32,32,239,239,239,239,96,96,96,96,0,0,0,0,239,239,239,239,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,96,96,96,96,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,143,143,143,143,255,255,255,255,255,255,255,255,255,255,255,255,239,239,239,239,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,239,239,239,239,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,64,64,64,64,80,80,80,80,64,64,64,64,48,48,48,48,0,0,0,0,0,0,0,0,
-	96,96,96,96,191,191,191,191,0,0,0,0,0,0,0,0,191,191,191,191,239,239,239,239,239,239,239,239,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,0,0,0,0,
-	128,128,128,128,255,255,255,255,112,112,112,112,16,16,16,16,0,0,0,0,112,112,112,112,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,143,143,143,143,255,255,255,255,255,255,255,255,255,255,255,255,80,80,80,80,223,223,223,223,255,255,255,255,255,255,255,255,191,191,191,191,96,96,96,96,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,207,207,207,207,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,239,239,239,239,64,64,64,64,255,255,255,255,255,255,255,255,255,255,255,255,207,207,207,207,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,96,96,96,96,191,191,191,191,255,255,255,255,255,255,255,255,223,223,223,223,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,96,96,96,96,
-	48,48,48,48,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,128,128,128,128,128,128,128,128,32,32,32,32,32,32,32,32,128,128,128,128,128,128,128,128,112,112,112,112,0,0,0,0,96,96,96,96,128,128,128,128,128,128,128,128,64,64,64,64,32,32,32,32,128,128,128,128,128,128,128,128,112,112,112,112,16,16,16,16,0,0,0,0,80,80,80,80,128,128,128,128,128,128,128,128,64,64,64,64,16,16,16,16,128,128,128,128,128,128,128,128,128,128,128,128,32,32,32,32,64,64,64,64,128,128,128,128,128,128,128,128,96,96,96,96,0,0,0,0,112,112,112,112,128,128,128,128,128,128,128,128,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,112,112,112,112,255,255,255,255,128,128,128,128,48,48,48,48,16,16,16,16,159,159,159,159,175,175,175,175,
-	0,0,0,0,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,207,207,207,207,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,255,255,255,255,255,255,255,255,239,239,239,239,96,96,96,96,255,255,255,255,255,255,255,255,255,255,255,255,143,143,143,143,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,207,207,207,207,80,80,80,80,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,207,207,207,207,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,143,143,143,143,255,255,255,255,255,255,255,255,255,255,255,255,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,112,112,112,112,175,175,175,175,48,48,48,48,48,48,48,48,0,0,0,0,80,80,80,80,191,191,191,191,
-	0,0,0,0,0,0,0,0,48,48,48,48,128,128,128,128,80,80,80,80,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,255,255,255,255,255,255,255,255,223,223,223,223,0,0,0,0,159,159,159,159,191,191,191,191,
-	191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,143,143,143,143,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,96,96,96,96,239,239,239,239,255,255,255,255,255,255,255,255,191,191,191,191,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,223,223,223,223,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,80,80,80,80,255,255,255,255,255,255,255,255,255,255,255,255,223,223,223,223,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,143,143,143,143,191,191,191,191,255,255,255,255,255,255,255,255,239,239,239,239,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,223,223,223,223,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,0,0,0,0,
-	255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,64,64,64,64,64,64,64,64,32,32,32,32,0,0,0,0,64,64,64,64,128,128,128,128,128,128,128,128,80,80,80,80,32,32,32,32,96,96,96,96,128,128,128,128,112,112,112,112,48,48,48,48,48,48,48,48,128,128,128,128,128,128,128,128,80,80,80,80,48,48,48,48,80,80,80,80,128,128,128,128,128,128,128,128,64,64,64,64,32,32,32,32,96,96,96,96,128,128,128,128,112,112,112,112,32,32,32,32,64,64,64,64,128,128,128,128,128,128,128,128,80,80,80,80,48,48,48,48,80,80,80,80,128,128,128,128,128,128,128,128,48,48,48,48,48,48,48,48,112,112,112,112,128,128,128,128,96,96,96,96,16,16,16,16,16,16,16,16,64,64,64,64,64,64,64,64,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,80,80,80,80,112,112,112,112,0,0,0,0,0,0,0,0,32,32,32,32,0,0,0,0,
-	0,0,0,0,96,96,96,96,239,239,239,239,255,255,255,255,255,255,255,255,96,96,96,96,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,80,80,80,80,255,255,255,255,255,255,255,255,255,255,255,255,207,207,207,207,80,80,80,80,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,191,191,191,191,255,255,255,255,255,255,255,255,239,239,239,239,96,96,96,96,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,64,64,64,64,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,96,96,96,96,239,239,239,239,255,255,255,255,255,255,255,255,191,191,191,191,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,80,80,80,80,223,223,223,223,255,255,255,255,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,0,0,0,0,207,207,207,207,223,223,223,223,48,48,48,48,239,239,239,239,255,255,255,255,143,143,143,143,
-	143,143,143,143,255,255,255,255,255,255,255,255,239,239,239,239,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,16,16,16,16,223,223,223,223,255,255,255,255,255,255,255,255,255,255,255,255,
-	255,255,255,255,255,255,255,255,255,255,255,255,207,207,207,207,191,191,191,191,143,143,143,143,0,0,0,0,0,0,0,0,0,0,0,0,96,96,96,96,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,207,207,207,207,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,143,143,143,143,239,239,239,239,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,239,239,239,239,143,143,143,143,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,64,64,64,64,223,223,223,223,
-	255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,96,96,96,96,128,128,128,128,128,128,128,128,64,64,64,64,32,32,32,32,128,128,128,128,128,128,128,128,128,128,128,128,32,32,32,32,80,80,80,80,128,128,128,128,128,128,128,128,96,96,96,96,16,16,16,16,112,112,112,112,128,128,128,128,128,128,128,128,64,64,64,64,64,64,64,64,128,128,128,128,128,128,128,128,128,128,128,128,64,64,64,64,128,128,128,128,128,128,128,128,112,112,112,112,0,0,0,0,128,128,128,128,128,128,128,128,128,128,128,128,64,64,64,64,64,64,64,64,128,128,128,128,128,128,128,128,112,112,112,112,16,16,16,16,96,96,96,96,128,128,128,128,128,128,128,128,64,64,64,64,32,32,32,32,128,128,128,128,128,128,128,128,128,128,128,128,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,175,175,175,175,255,255,255,255,255,255,255,255,191,191,191,191,96,96,96,96,239,239,239,239,
-	0,0,0,0,0,0,0,0,32,32,32,32,64,64,64,64,128,128,128,128,159,159,159,159,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,239,239,239,239,48,48,48,48,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,64,64,64,64,255,255,255,255,255,255,255,255,239,239,239,239,96,96,96,96,191,191,191,191,255,255,255,255,255,255,255,255,207,207,207,207,80,80,80,80,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,48,48,48,48,223,223,223,223,255,255,255,255,255,255,255,255,128,128,128,128,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,80,80,80,80,207,207,207,207,255,255,255,255,255,255,255,255,191,191,191,191,96,96,96,96,255,255,255,255,255,255,255,255,255,255,255,255,80,80,80,80,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,0,0,0,0,80,80,80,80,112,112,112,112,112,112,112,112,16,16,16,16,32,32,32,32,80,80,80,80,
-	143,143,143,143,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,64,64,64,64,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,191,191,191,191,128,128,128,128,64,64,64,64,0,0,0,0,
-	175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,239,239,239,239,191,191,191,191,64,64,64,64,0,0,0,0,0,0,0,0,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,223,223,223,223,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,223,223,223,223,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,
-	64,64,64,64,112,112,112,112,64,64,64,64,16,16,16,16,0,0,0,0,96,96,96,96,143,143,143,143,0,0,0,0,0,0,0,0,16,16,16,16,159,159,159,159,191,191,191,191,191,191,191,191,96,96,96,96,48,48,48,48,191,191,191,191,191,191,191,191,191,191,191,191,48,48,48,48,128,128,128,128,191,191,191,191,191,191,191,191,143,143,143,143,32,32,32,32,175,175,175,175,191,191,191,191,191,191,191,191,96,96,96,96,80,80,80,80,191,191,191,191,191,191,191,191,191,191,191,191,96,96,96,96,191,191,191,191,191,191,191,191,159,159,159,159,16,16,16,16,191,191,191,191,191,191,191,191,191,191,191,191,80,80,80,80,96,96,96,96,191,191,191,191,191,191,191,191,175,175,175,175,32,32,32,32,143,143,143,143,191,191,191,191,191,191,191,191,128,128,128,128,48,48,48,48,191,191,191,191,191,191,191,191,175,175,175,175,48,48,48,48,0,0,0,0,0,0,0,0,80,80,80,80,223,223,223,223,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,
-	16,16,16,16,16,16,16,16,0,0,0,0,0,0,0,0,32,32,32,32,239,239,239,239,223,223,223,223,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,191,191,191,191,191,191,191,191,191,191,191,191,112,112,112,112,48,48,48,48,191,191,191,191,191,191,191,191,175,175,175,175,48,48,48,48,143,143,143,143,191,191,191,191,191,191,191,191,143,143,143,143,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,80,80,80,80,191,191,191,191,191,191,191,191,191,191,191,191,48,48,48,48,143,143,143,143,191,191,191,191,191,191,191,191,143,143,143,143,48,48,48,48,191,191,191,191,191,191,191,191,191,191,191,191,48,48,48,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,255,255,255,255,255,255,255,255,239,239,239,239,223,223,223,223,64,64,64,64,16,16,16,16,
-	32,32,32,32,239,239,239,239,128,128,128,128,32,32,32,32,191,191,191,191,255,255,255,255,223,223,223,223,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,112,112,112,112,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,223,223,223,223,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,239,239,239,239,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,239,239,239,239,159,159,159,159,32,32,32,32,32,32,32,32,239,239,239,239,175,175,175,175,0,0,0,0,
-	0,0,0,0,159,159,159,159,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,223,223,223,223,159,159,159,159,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,239,239,239,239,255,255,255,255,255,255,255,255,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,80,80,80,80,64,64,64,64,16,16,16,16,0,0,0,0,128,128,128,128,175,175,175,175,0,0,0,0,
-	0,0,0,0,16,16,16,16,112,112,112,112,239,239,239,239,255,255,255,255,255,255,255,255,255,255,255,255,239,239,239,239,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,96,96,96,96,255,255,255,255,255,255,255,255,255,255,255,255,80,80,80,80,191,191,191,191,255,255,255,255,255,255,255,255,191,191,191,191,64,64,64,64,239,239,239,239,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,143,143,143,143,128,128,128,128,255,255,255,255,255,255,255,255,239,239,239,239,80,80,80,80,191,191,191,191,255,255,255,255,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,207,207,207,207,255,255,255,255,255,255,255,255,239,239,239,239,96,96,96,96,0,0,0,0,64,64,64,64,0,0,0,0,
-	0,0,0,0,16,16,16,16,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,207,207,207,207,159,159,159,159,80,80,80,80,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,191,191,191,191,223,223,223,223,207,207,207,207,255,255,255,255,255,255,255,255,255,255,255,255,32,32,32,32,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,175,175,175,175,207,207,207,207,112,112,112,112,48,48,48,48,16,16,16,16,80,80,80,80,239,239,239,239,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,207,207,207,207,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,255,255,255,255,16,16,16,16,0,0,0,0,64,64,64,64,207,207,207,207,255,255,255,255,64,64,64,64,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,159,159,159,159,175,175,175,175,16,16,16,16,239,239,239,239,143,143,143,143,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,239,239,239,239,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,159,159,159,159,0,0,0,0,0,0,0,0,16,16,16,16,255,255,255,255,32,32,32,32,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,159,159,159,159,128,128,128,128,32,32,32,32,255,255,255,255,239,239,239,239,255,255,255,255,255,255,255,255,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,223,223,223,223,112,112,112,112,143,143,143,143,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,207,207,207,207,255,255,255,255,255,255,255,255,255,255,255,255,175,175,175,175,175,175,175,175,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,207,207,207,207,48,48,48,48,191,191,191,191,255,255,255,255,255,255,255,255,255,255,255,255,239,239,239,239,48,48,48,48,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,159,159,159,159,255,255,255,255,255,255,255,255,191,191,191,191,48,48,48,48,16,16,16,16,207,207,207,207,175,175,175,175,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,207,207,207,207,255,255,255,255,255,255,255,255,96,96,96,96,96,96,96,96,175,175,175,175,159,159,159,159,48,48,48,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,207,207,207,207,96,96,96,96,0,0,0,0,0,0,0,0,0,0,0,0,96,96,96,96,255,255,255,255,143,143,143,143,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,159,159,159,159,207,207,207,207,239,239,239,239,255,255,255,255,255,255,255,255,159,159,159,159,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,112,112,112,112,255,255,255,255,255,255,255,255,239,239,239,239,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,80,80,80,80,191,191,191,191,48,48,48,48,48,48,48,48,207,207,207,207,255,255,255,255,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,159,159,159,159,255,255,255,255,255,255,255,255,239,239,239,239,112,112,112,112,128,128,128,128,96,96,96,96,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,80,80,80,80,255,255,255,255,80,80,80,80,0,0,0,0,0,0,0,0,16,16,16,16,175,175,175,175,255,255,255,255,143,143,143,143,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,143,143,143,143,255,255,255,255,255,255,255,255,207,207,207,207,32,32,32,32,239,239,239,239,255,255,255,255,255,255,255,255,175,175,175,175,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,255,255,255,255,159,159,159,159,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,96,96,96,96,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,239,239,239,239,175,175,175,175,16,16,16,16,64,64,64,64,239,239,239,239,255,255,255,255,143,143,143,143,255,255,255,255,207,207,207,207,48,48,48,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,223,223,223,223,143,143,143,143,112,112,112,112,255,255,255,255,239,239,239,239,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,0,0,0,0,96,96,96,96,255,255,255,255,255,255,255,255,239,239,239,239,207,207,207,207,255,255,255,255,255,255,255,255,64,64,64,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,207,207,207,207,255,255,255,255,255,255,255,255,255,255,255,255,223,223,223,223,255,255,255,255,80,80,80,80,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,112,112,112,112,255,255,255,255,175,175,175,175,207,207,207,207,255,255,255,255,255,255,255,255,112,112,112,112,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,128,128,128,128,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,159,159,159,159,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,207,207,207,207,223,223,223,223,255,255,255,255,255,255,255,255,159,159,159,159,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,191,191,191,191,255,255,255,255,255,255,255,255,159,159,159,159,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,16,16,16,16,159,159,159,159,255,255,255,255,207,207,207,207,16,16,16,16,0,0,0,0,0,0,0,0,143,143,143,143,239,239,239,239,80,80,80,80,0,0,0,0,0,0,0,0,48,48,48,48,191,191,191,191,128,128,128,128,96,96,96,96,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,0,0,0,0,112,112,112,112,191,191,191,191,175,175,175,175,0,0,0,0,0,0,0,0,16,16,16,16,223,223,223,223,223,223,223,223,16,16,16,16,0,0,0,0,0,0,0,0,0,0,0,0,64,64,64,64,143,143,143,143,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,96,96,96,96,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,223,223,223,223,255,255,255,255,112,112,112,112,0,0,0,0,0,0,0,0,96,96,96,96,255,255,255,255,255,255,255,255,207,207,207,207,255,255,255,255,112,112,112,112,143,143,143,143,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,191,143,143,143,143,239,239,239,239,255,255,255,255,48,48,48,48,128,128,128,128,255,255,255,255,223,223,223,223,0,0,0,0,0,0,0,0,48,48,48,48,255,255,255,255,255,255,255,255,48,48,48,48,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,112,112,112,112,16,16,16,16,0,0,0,0,0,0,0,0,128,128,128,128,255,255,255,255,207,207,207,207,159,159,159,159,255,255,255,255,191,191,191,191,48,48,48,48,112,112,112,112,255,255,255,255,255,255,255,255,64,64,64,64,32,32,32,32,255,255,255,255,255,255,255,255,96,96,96,96,112,112,112,112,255,255,255,255,255,255,255,255,32,32,32,32,0,0,0,0,0,0,0,0,96,96,96,96,96,96,96,96,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,191,191,191,191,255,255,255,255,239,239,239,239,255,255,255,255,239,239,239,239,48,48,48,48,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,207,207,207,207,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,255,80,80,80,80,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,239,239,239,239,255,255,255,255,128,128,128,128,255,255,255,255,255,255,255,255,96,96,96,96,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,159,159,159,159,255,255,255,255,191,191,191,191,0,0,0,0,223,223,223,223,255,255,255,255,128,128,128,128,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,32,32,32,32,255,255,255,255,255,255,255,255,175,175,175,175,255,255,255,255,255,255,255,255,96,96,96,96,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,112,112,112,112,255,255,255,255,255,255,255,255,0,0,0,0,175,175,175,175,255,255,255,255,191,191,191,191,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-	0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,48,48,48,48,80,80,80,80,128,128,128,128,128,128,128,128,96,96,96,96,0,0,0,0,0,0,0,0,64,64,64,64,255,255,255,255,255,255,255,255,0,0,0,0,0,0,0,0,48,48,48,48,255,255,255,255,191,191,191,191,16,16,16,16,32,32,32,32,32,32,32,32,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-};
